+++ maint/gen_binding_f08.py
@@ -25,7 +25,6 @@ def main():
     else:
         # FIXME: until romio interface is generated
         func_list.extend(get_mpiio_func_list())
-    func_list.extend(get_type_create_f90_func_list())
     func_list.append(G.FUNCS['mpi_f_sync_reg'])
 
     # preprocess
