+++ maint/gen_binding_f90.py
@@ -20,7 +20,6 @@ def main():
         func_list = [f for f in func_list if not f['name'].startswith('MPI_File_')]
     else:
         func_list.extend(get_mpiio_func_list())
-    func_list.extend(get_type_create_f90_func_list())
     func_list.extend(get_f77_dummy_func_list())
 
     def has_cptr(func):
