--- src/binding/fortran/use_mpi/create_f90_util.h
@@ -1,15 +1,0 @@
-/*
- * Copyright (C) by Argonne National Laboratory
- *     See COPYRIGHT in top-level directory
- */
-
-#ifndef CREATE_F90_UTIL_H_INCLUDED
-#define CREATE_F90_UTIL_H_INCLUDED
-
-#include "mpiimpl.h"
-
-int MPIR_Create_unnamed_predefined(MPI_Datatype old, int combiner,
-                                   int r, int p, MPI_Datatype * new_ptr);
-
-
-#endif /* CREATE_F90_UTIL_H_INCLUDED */
